refactor(mllp-client)!: drop Uint8Array — speak string + Root only; split hl7v2.ts
Narrow the public API to logical-message representations:

- SendInput is now `string | Root` (was `string | Uint8Array | Root`).
  Uint8Array was the only input whose type name implied byte-fidelity — a
  contract the cleaning round-trip cannot keep — while it was merely
  decoded-then-parsed identically to a string. Callers holding bytes decode at
  their I/O boundary (where charset/MSH-18 lives) and pass the string.
- MllpClientResponse.raw and MllpClientError.raw are now `string` (decoded ACK
  text), matching @glion/ack AckException.raw. Both directions speak string, so
  receive-then-forward no longer needs a hand-decode. The send-side
  PARSE_FAILED-on-non-UTF-8 case is gone; PARSE_FAILED still surfaces for a
  non-UTF-8 inbound ACK (strict UTF-8 decode).

Reorg: retire the format-named hl7v2.ts (meaningless in the client) and split
it along the send/response seam into send.ts (SendInput, PreparedSend,
prepareSend) and response.ts (MllpClientResponse, parseResponse). No public
entry-point change.

CLI consumer (@glion/cli) passes a string and never reads response.raw, so it
is unaffected (type-checks clean). 96 tests pass.
