+++ packages/mllp-client/README.md
@@ -4,7 +4,7 @@ Persistent, single-flight MLLP client for HL7v2.
 
 ## What it does
 
-`@glion/mllp-client` opens one long-lived MLLP/TCP connection and sends HL7v2 messages over it, one in flight at a time. `send()` accepts a `string`, `Uint8Array`, or parsed `Root` AST. The AST is the first-class currency: every input is parsed to a tree and re-serialized to **canonical** HL7v2 for the wire — this is an _originating / cleaning_ client, not a byte-exact relay. Cleaning is syntactic only (line endings normalized to CR, trailing empty fields/segments trimmed); escape sequences, Z-segments, repetitions, and components are preserved. The same parse drives MSH-10 correlation. Each `send()` waits for the peer's ACK, verifies the MSA-2 correlation ID against the request's MSH-10, and resolves with a structured response. Application/commit accept codes (`AA`/`CA`) resolve; reject codes (`AE`/`AR`/`CE`/`CR`) throw the matching `@glion/ack` `AckException` (`AckApplicationError`, `AckApplicationReject`, `AckCommitError`, `AckCommitReject`).
+`@glion/mllp-client` opens one long-lived MLLP/TCP connection and sends HL7v2 messages over it, one in flight at a time. `send()` accepts a `string` or a parsed `Root` AST. The AST is the first-class currency: every input is parsed to a tree and re-serialized to **canonical** HL7v2 for the wire — this is an _originating / cleaning_ client, not a byte-exact relay. Raw bytes are not accepted: decode them to text at your I/O boundary (where charset / MSH-18 knowledge lives) and pass the `string`. Cleaning is syntactic only (line endings normalized to CR, trailing empty fields/segments trimmed); escape sequences, Z-segments, repetitions, and components are preserved. The same parse drives MSH-10 correlation. Each `send()` waits for the peer's ACK, verifies the MSA-2 correlation ID against the request's MSH-10, and resolves with a structured response. Application/commit accept codes (`AA`/`CA`) resolve; reject codes (`AE`/`AR`/`CE`/`CR`) throw the matching `@glion/ack` `AckException` (`AckApplicationError`, `AckApplicationReject`, `AckCommitError`, `AckCommitReject`).
 
 The client is one class managing one socket lifecycle: `connect()` initializes, `send()` uses, `close()` tears down. Concurrent sends queue (FIFO) and run one at a time. The wire transport is supplied by a runtime adapter — the Node adapter ships at `@glion/mllp-client/node`. A peer drop, write failure, or decoder framing error is terminal: the connection moves to `closed` and the instance is spent. Reconnect is deferred to a later version.
 
@@ -94,7 +94,7 @@ All are an `MllpClientError`; branch on `code`:
 
 ### `client.send(message, options?): Promise<MllpClientResponse>`
 
-Accepts a `string`, `Uint8Array`, or `Root` (`SendInput`). All inputs are parsed to a tree and re-serialized to **canonical** HL7v2 for the wire (a `string`/`Uint8Array` is parsed first; a `Root` is used directly), and the same tree yields the MSH-10 correlation ID. The wire bytes are the cleaned form, not the caller's original bytes — line endings normalize to CR and trailing empty fields/segments are trimmed, while escape sequences and structure are preserved. (Known limitations: trailing-empty trimming is not idempotent; a `Root` that was escape-_decoded_ upstream must not be passed in, as it would re-serialize the decoded literals; non-UTF-8 input bytes are rejected with `PARSE_FAILED`.) `options.timeoutMs` overrides the default ACK-wait deadline; the deadline starts when the send reaches the wire (a send waiting behind others may wait longer than this before its clock begins). There is no per-send cancellation signal — `close()` rejects an in-flight or queued send.
+Accepts a `string` or a `Root` (`SendInput`) — raw bytes are not accepted; decode them to text at your I/O boundary (where charset / MSH-18 knowledge lives) and pass the `string`. Both inputs are parsed to a tree and re-serialized to **canonical** HL7v2 for the wire (a `string` is parsed; a `Root` is used directly), and the same tree yields the MSH-10 correlation ID. The wire bytes are the cleaned form, not a byte-exact echo of the input — line endings normalize to CR and trailing empty fields/segments are trimmed, while escape sequences and structure are preserved. (Known limitations: trailing-empty trimming is not idempotent; a `Root` that was escape-_decoded_ upstream must not be passed in, as it would re-serialize the decoded literals.) `options.timeoutMs` overrides the default ACK-wait deadline; the deadline starts when the send reaches the wire (a send waiting behind others may wait longer than this before its clock begins). There is no per-send cancellation signal — `close()` rejects an in-flight or queued send.
 
 **Throws**
 
@@ -105,7 +105,7 @@ Accepts a `string`, `Uint8Array`, or `Root` (`SendInput`). All inputs are parsed
   - `SEND_TIMEOUT` — no ACK arrived within the deadline (`timeoutMs` is set).
   - `DROPPED` — the connection ended (peer drop, write failure, framing error, frame flood); `reason` discriminates. Terminal.
   - `NOT_CONNECTED` / `CLOSED` — the client is not connected, or has been closed (the in-flight send and every queued send reject).
-  - `PARSE_FAILED` — the ACK is not parseable HL7v2, or the request bytes are not valid UTF-8 (the request is parsed before sending, so a non-UTF-8 `Uint8Array` is rejected here).
+  - `PARSE_FAILED` — the peer's ACK is not parseable HL7v2, or is not valid UTF-8 (a Latin-1 / Windows-1252 peer ACK trips this; the client decodes ACKs as strict UTF-8).
   - `UNKNOWN_ACK_CODE` — the ACK carries a non-standard MSA-1.
 
 ### `client.close(): Promise<void>`
@@ -124,7 +124,7 @@ Calls `close()`. Enables `await using`.
 | `controlId`        | `string`         | MSA-2 echoed by the peer; `""` when the peer omits it.                    |
 | `requestControlId` | `string`         | MSH-10 of the request this ACK answers.                                   |
 | `tree`             | `Root`           | Parsed AST of the ACK, for arbitrary field access via `value()`.          |
-| `raw`              | `Uint8Array`     | De-framed ACK payload bytes.                                              |
+| `raw`              | `string`         | De-framed ACK payload as decoded text (UTF-8).                            |
 | `timestamp`        | `Date`           | Wall-clock instant the ACK finished arriving.                             |
 | `durationMs`       | `number`         | Wire-level round-trip, measured monotonically.                            |
