+++ packages/mllp-client/src/index.ts
@@ -1,10 +1,11 @@
 /**
  * `@glion/mllp-client` — persistent MLLP client for HL7v2.
  *
- * One connection, one send on the wire at a time. `send()` accepts a
- * `string` / `Uint8Array` / `Root`: the caller's bytes (or a serialized
- * `Root`) go on the wire unchanged, while the tree is used to read MSH-10 for
- * correlation and to parse the ACK. A NAK (AE/AR/CE/CR) throws the central
+ * One connection, one send on the wire at a time. `send()` accepts a `string`
+ * or a `Root`: every input is parsed to a tree and re-serialized to canonical
+ * HL7v2 for the wire (an *originating / cleaning* client, not a byte-exact
+ * relay), and the same tree reads MSH-10 for correlation. A NAK (AE/AR/CE/CR)
+ * throws the central
  * `@glion/ack` exception family (`AckApplicationError`, `AckApplicationReject`,
  * `AckCommitError`, `AckCommitReject`) — the same types the server throws;
  * import them and `AckException` from `@glion/ack`. Runtime adapters live
@@ -23,4 +24,5 @@ export type {
 export type { MllpConnector, MllpDuplex } from "./duplex";
 export { MllpClientError, MllpErrorCode } from "./errors";
 export type { MllpClientErrorDetails, MllpDropReason } from "./errors";
-export type { MllpClientResponse, SendInput } from "./hl7v2";
+export type { MllpClientResponse } from "./response";
+export type { SendInput } from "./send";
