+++ packages/mllp-client/src/response.ts
@@ -1,32 +1,11 @@
 /**
- * HL7v2 operations for the client. The AST (`Root`) is the first-class send
- * currency: every `send()` input is normalized to a tree and the wire bytes are
- * produced from that tree with `@glion/to-hl7v2`. The client is therefore an
- * *originating / cleaning* client — it emits canonical HL7v2, not a byte-exact
- * relay of whatever arrived. A `string` / `Uint8Array` is parsed first (it is a
- * serialized tree); a `Root` is used directly.
- *
- * **What "cleaning" changes** (syntactic only — semantics are preserved):
- * line endings are normalized to CR, and trailing empty fields / segments are
- * trimmed. Escape sequences (`\F\`, `\X0D\`, …), Z-segments, repetitions, and
- * components are preserved verbatim by the parser→serializer round trip.
- *
- * **Known limitations** (documented, accepted):
- *
- * - Trailing-empty-field trimming is NOT idempotent — it drops one trailing empty
- *   field per pass (`PID|1|||||` → `PID|1||||` → … → `PID|1`), so cleaning the
- *   same message twice can yield different bytes. Semantically faithful (the
- *   fields are absent either way); syntactically non-convergent.
- * - **Decode-implies-encode invariant:** the client parses with the base
- *   `@glion/parser`, which does NOT decode escape sequences — so `\F\` in →
- *   `\F\` out, no re-encode needed. Do NOT hand `send()` a `Root` that has
- *   already been escape-decoded (e.g. via `hl7v2DecodeEscapes`): `toHl7v2` has
- *   no re-encode step and would emit the decoded literal, corrupting field
- *   structure.
- * - Non-UTF-8 input bytes (Latin-1 / Windows-1252) cannot be decoded and are
- *   rejected. This inherits the glion ecosystem's UTF-8 assumption; whether
- *   that assumption is correct is tracked separately (see the encoding GH
- *   issue), not papered over here.
+ * Response-side ACK parsing. Turns the peer's de-framed ACK bytes into a
+ * structured {@link MllpClientResponse}, or throws. The ACK is decoded as
+ * strict UTF-8 (a non-UTF-8 peer is rejected with `PARSE_FAILED` rather than
+ * silently substituted — this inherits the glion ecosystem's UTF-8 assumption,
+ * tracked separately via the encoding GH issue), parsed, and correlated against
+ * the request's MSH-10. Accept codes (`AA`/`CA`) resolve; a NAK (`AE`/`AR`/
+ * `CE`/`CR`) throws the matching `@glion/ack` `AckException`.
  *
  * @module
  */
@@ -40,9 +19,7 @@ import {
 } from "@glion/ack";
 import type { AckSuccessCode } from "@glion/ack";
 import type { Root } from "@glion/ast";
-import { frame } from "@glion/mllp-transport";
 import { parseHL7v2 } from "@glion/parser";
-import { toHl7v2 } from "@glion/to-hl7v2";
 import { value } from "@glion/util-query";
 
 import { MllpClientError, MllpErrorCode } from "./errors";
@@ -53,13 +30,6 @@ import { MllpClientError, MllpErrorCode } from "./errors";
  * fail PARSE_FAILED rather than silently substitute U+FFFD.
  */
 const TEXT_DECODER = new TextDecoder("utf-8", { fatal: true });
-
-/**
- * What `MllpClient.send()` accepts. All inputs are normalized to a tree and
- * serialized to canonical HL7v2 for the wire (see the module JSDoc): a `string`
- * / `Uint8Array` is parsed first; a `Root` is used directly.
- */
-export type SendInput = string | Uint8Array | Root;
 
 export interface MllpClientResponse {
   /**
@@ -76,66 +46,12 @@ export interface MllpClientResponse {
   readonly requestControlId: string;
   /** Parsed AST of the ACK message. */
   readonly tree: Root;
-  /** De-framed payload bytes. */
-  readonly raw: Uint8Array;
+  /** De-framed ACK payload as decoded text (UTF-8). */
+  readonly raw: string;
   /** Wall-clock instant the ACK frame finished arriving. */
   readonly timestamp: Date;
   /** Wire-level round-trip duration (monotonic), milliseconds. */
   readonly durationMs: number;
-}
-
-/** What a send needs on the wire, derived from a single parse of the input. */
-export interface PreparedSend {
-  /** Canonical HL7v2 wire bytes, MLLP-framed. */
-  readonly framed: Uint8Array;
-  /** MSH-10 of the (cleaned) message, for ACK correlation. `""` if absent. */
-  readonly requestControlId: string;
-}
-
-/**
- * Normalize a send input to its canonical wire form and read its control ID —
- * from ONE parse. A `string` / `Uint8Array` is parsed to a tree (the bytes are
- * decoded as UTF-8 first); a `Root` is used directly. The tree is then
- * serialized with `@glion/to-hl7v2` and MLLP-framed, and MSH-10 is read from
- * the same tree. The wire bytes are therefore the *cleaned* canonical form, not
- * the caller's original bytes — see the module JSDoc.
- *
- * @throws {MllpClientError} `PARSE_FAILED` when a `Uint8Array` is not valid
- *   UTF-8 (e.g. a Latin-1 / Windows-1252 feed). The cause carries the decode
- *   error. This inherits the ecosystem's UTF-8 assumption.
- * @throws {FramingError} When the serialized message carries an embedded MLLP
- *   framing byte (VT or FS) that cannot be framed. CR is allowed — it is the
- *   HL7v2 segment terminator.
- */
-export function prepareSend(input: SendInput): PreparedSend {
-  const tree = toTree(input);
-  return {
-    framed: frame(toHl7v2(tree)),
-    requestControlId: readValue(tree, "MSH-10[1].1.1") ?? "",
-  };
-}
-
-function toTree(input: SendInput): Root {
-  if (typeof input !== "string" && !(input instanceof Uint8Array)) {
-    return input;
-  }
-  let text: string;
-  if (typeof input === "string") {
-    text = input;
-  } else {
-    try {
-      text = TEXT_DECODER.decode(input);
-    } catch (error) {
-      throw new MllpClientError(
-        MllpErrorCode.PARSE_FAILED,
-        "The message bytes are not valid UTF-8 and cannot be parsed (HL7v2 in the glion ecosystem is UTF-8; a Latin-1 / Windows-1252 feed is not yet supported).",
-        { cause: error }
-      );
-    }
-  }
-  // The parser is lenient — it never throws — so a tree is always produced,
-  // even for non-HL7v2 text (MSH-10 then reads as "").
-  return parseHL7v2(text);
 }
 
 interface ParseInput {
@@ -154,8 +70,7 @@ export function parseResponse(input: ParseInput): MllpClientResponse {
   } catch (error) {
     // Strict UTF-8 decode (fatal): a Latin-1 / Windows-1252 peer must surface
     // as PARSE_FAILED, not a raw TypeError, so the contract "every failure is
-    // an MllpClientError you can branch on by `code`" holds on the ACK path
-    // just as the request side (readableTree) already guards it.
+    // an MllpClientError you can branch on by `code`" holds on the ACK path.
     throw new MllpClientError(
       MllpErrorCode.PARSE_FAILED,
       "The peer's ACK is not valid UTF-8; HL7v2 messages must be ASCII/UTF-8 (a Latin-1 / Windows-1252 peer trips this). See the error's cause.",
@@ -201,7 +116,7 @@ export function parseResponse(input: ParseInput): MllpClientResponse {
     throw new MllpClientError(
       MllpErrorCode.CORRELATION_MISMATCH,
       `ACK control-ID mismatch: the response's MSA-2 ("${controlId}") does not match the request's MSH-10 ("${requestControlId}"). This usually means a late ACK from a previously-timed-out request arrived on this connection.`,
-      { actual: controlId, expected: requestControlId, raw, tree }
+      { actual: controlId, expected: requestControlId, raw: text, tree }
     );
   }
 
@@ -244,7 +159,7 @@ export function parseResponse(input: ParseInput): MllpClientResponse {
     code: codeRaw,
     controlId,
     durationMs,
-    raw,
+    raw: text,
     requestControlId,
     timestamp,
     tree,
